pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.118.0",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
